refactor: move scrolling features to scrolling package
